setPixelColor replaced with drawPixel
NeoMatrix (24 bit colors) inherits GFX (16 bit colors), so we use
drawPixel to bypass GFX limitations
Included antiDelay from esp8266 branch, with a modification for it to
return 1 on exit or 0 on new web data which would end whatever function
is currently running (stops p_cylon, etc).
Per esp8266, also removed Snake and MQTT. Looking in to a different MQTT
library that may be added once esp8266 research is done.
line 870 is a debugging check that the main loop is still running/not
stuck
line 3 sets Webduino library to debug connections to Serial