Update Header, new Javascript color picker
Now using jscolor-2.0.4, function at bottom is jscolor.min.js from that
release.
Updated header file, spaced out the html a bit better.
All functions that use a selected color now have a button to do color
picking with. This translates to an RGB value that gets sent with the
function. Looking in to a way to clean up the separate 'send' functions
as well.
To serve the JS file, a new cmd_js was added. http://tableip/color.js
will cause the Javascript to get sent. It has a cache-time of 30 days,
shouldn't need to be updated once it gets requested.